etl/reddit_collector/requirements.txt
pymongo
requests
dnspython
python-dotenv
praw